+++ AddOns/Platynator/CustomiseDialog/WidgetsConfiguration.lua
@@ -796,6 +796,23 @@ addonTable.CustomiseDialog.WidgetsConfig = {
         }
       }
     },
+    ["castSpellName"] = {
+      {
+        label = addonTable.Locales.GENERAL,
+        entries = {
+          {
+            label = addonTable.Locales.SHOW_ON_INTERRUPT,
+            kind = "checkbox",
+            setter = function(details, value)
+              details.showInterrupted = value
+            end,
+            getter = function(details)
+              return details.showInterrupted
+            end,
+          },
+        }
+      }
+    },
     ["castTarget"] = {
       {
         label = addonTable.Locales.COLORS,